Other/api integrate/requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
pillow
pandas
python-multipart
pydantic
